added duke package
