+++ src/main/java/duke/ByeCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList taskList, Ui ui) {
