+++ src/main/java/duke/Duke.java
@@ -1,17 +1,13 @@
-import java.time.LocalDate;
+package duke;
+
 import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Scanner;
 
 public class Duke {
-    private Storage storage;
     private TaskList tasks;
     private Ui ui;
 
     public Duke() {
         ui = new Ui();
-        storage = new Storage();
         tasks = new TaskList();
     }
 
@@ -21,7 +17,7 @@ public class Duke {
             + "| |_| | |_| |   <  __/\n"
             + "|____/ \\__,_|_|\\_\\___|\n";
 
-    private static final String GREETING = "Hello! I'm Duke\n"
+    private static final String GREETING = "Hello! I'm duke.Duke\n"
             + "What can I do for you?\n";
 
     public static void main(String[] args) {
