+++ src/main/java/duke/DukeInvalidDescriptionException.java
@@ -1,5 +1,7 @@
+package duke;
+
 /**
- * A DukeInvalidDescriptionException is thrown if the user tries to add a task without giving it either a name or
+ * A duke.DukeInvalidDescriptionException is thrown if the user tries to add a task without giving it either a name or
  * description, or if no due date or time is given for Events and Deadlines.
  */
 public class DukeInvalidDescriptionException extends DukeException {
@@ -7,7 +9,7 @@ public class DukeInvalidDescriptionException extends DukeException {
     /**
      * Public constructor for a DukeEmptyDescriptionException.
      *
-     * @param message the message to be printed when a DukeInvalidDescriptionException is thrown and caught.
+     * @param message the message to be printed when a duke.DukeInvalidDescriptionException is thrown and caught.
      */
     public DukeInvalidDescriptionException(String message) {
         super(message);
