+++ src/main/java/duke/EventCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 
 public class EventCommand extends Command {
