+++ src/main/java/duke/MarkCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class MarkCommand extends Command {
     private int index;
