+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 
 public abstract class Task {
