+++ src/main/java/duke/TaskType.java
@@ -1,3 +1,4 @@
+package duke;
 
 public enum TaskType {
     T,
