+++ text-ui-test/runtest.txt
@@ -15,7 +15,7 @@ IF ERRORLEVEL 1 (
 REM no error here, errorlevel == 0
 
 REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..C:\Users\ykaij\Desktop\documents\NUS\Y2S1\CS2103T\IP\src\main Duke < input.txt > ACTUAL.TXT
+java -classpath ..C:\Users\ykaij\Desktop\documents\NUS\Y2S1\CS2103T\IP\src\main duke.Duke < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
